update page, and update pagination method

+++ flowers_common/src/main/java/com/flowers/common/bean/ResultJson.java
@@ -6,12 +6,12 @@ public class ResultJson {
     private String code = "0";
     private Object body = null;
     private String msg = "成功";
-    private Integer count = null;
+    private Long count = null;
 
     public ResultJson() {
     }
 
-    public ResultJson(String code, Object body, String msg, Integer count) {
+    public ResultJson(String code, Object body, String msg, Long count) {
         this.code = code;
         this.body = body;
         this.msg = msg;
@@ -24,7 +24,7 @@ public class ResultJson {
         this.msg = msg;
     }
 
-    public ResultJson(String code, Object body, Integer count) {
+    public ResultJson(String code, Object body, Long count) {
         this.code = code;
         this.body = body;
         this.count = count;
@@ -35,7 +35,7 @@ public class ResultJson {
         this.body = body;
     }
 
-    public ResultJson(Object body, String msg, Integer count) {
+    public ResultJson(Object body, String msg, Long count) {
         this.body = body;
         this.msg = msg;
         this.count = count;
@@ -46,7 +46,7 @@ public class ResultJson {
         this.msg = msg;
     }
 
-    public ResultJson(Object body, Integer count) {
+    public ResultJson(Object body, Long count) {
         this.body = body;
         this.count = count;
     }
@@ -85,11 +85,11 @@ public class ResultJson {
         this.code = code;
     }
 
-    public Integer getCount() {
+    public Long getCount() {
         return count;
     }
 
-    public void setCount(Integer count) {
+    public void setCount(Long count) {
         this.count = count;
     }
 }
